requirements.txt
django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1